Adding deamonset status update using TPR - VcpStatus (#230)
* adding deamonset status update using TPR - VcpStatus

+++ enable-vcp-image/enable-vcp-scripts/daemonset_pod.sh
@@ -3,8 +3,16 @@ source $(dirname "$0")/common_func.sh
 source $(dirname "$0")/exit_codes.sh
 
 [ -z "$POD_NAME" ] && { echo "[ERROR] POD_NAME is not set"; exit $ERROR_POD_ENV_VALIDATION; }
-[ -z "$NODE_NAME" ] && { echo "[ERROR] NODE_NAME is not set"; exit $ERROR_POD_ENV_VALIDATION; }
-[ -z "$POD_ROLE" ] && { echo "[ERROR] POD_ROLE is not set"; exit $ERROR_POD_ENV_VALIDATION; }
+init_VcpConigStatus "$POD_NAME"
+
+PHASE=$DAEMONSET_SCRIPT_PHASE1
+update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_RUNNING" ""
+
+ERROR_MSG="NODE_NAME is not set"
+[ -z "$NODE_NAME" ] && {update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"; exit $ERROR_POD_ENV_VALIDATION; }
+
+ERROR_MSG="POD_ROLE is not set"
+[ -z "$POD_ROLE" ] && {update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"; exit $ERROR_POD_ENV_VALIDATION; }
 
 echo "Running script in the Pod:" $POD_NAME "deployed on the Node:" $NODE_NAME
 # read secret keys from volume /secret-volume/ and set values in an environment
@@ -16,33 +24,47 @@ else
     echo "Daemonset pod is running on the worker node"
 fi
 
+PHASE=$DAEMONSET_SCRIPT_PHASE2
+update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_RUNNING" ""
+
 # connect to vCenter using VC Admin username and password
 export GOVC_INSECURE=1
 export GOVC_URL='https://'$k8s_secret_vc_admin_username':'$k8s_secret_vc_admin_password'@'$k8s_secret_vc_ip':'$k8s_secret_vc_port'/sdk'
 
 # Get VM's UUID, Find VM Path using VM UUID and set disk.enableUUID to 1 on the VM
 vmuuid=$(cat /host/sys/class/dmi/id/product_serial | sed -e 's/^VMware-//' -e 's/-/ /' | awk '{ print tolower($1$2$3$4 "-" $5$6 "-" $7$8 "-" $9$10 "-" $11$12$13$14$15$16) }')
-[ -z "$vmuuid" ] && { echo "[ERROR] Unable to get VM UUID from /host/sys/class/dmi/id/product_serial"; exit $ERROR_UNKNOWN; }
+ERROR_MSG="Unable to get VM UUID from /host/sys/class/dmi/id/product_serial"
+[ -z "$vmuuid" ] && {update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"; exit $ERROR_UNKNOWN; }
 
 vmpath=$(govc vm.info -vm.uuid=$vmuuid | grep "Path:" | awk 'BEGIN {FS=":"};{print $2}' | tr -d ' ')
-[ -z "$vmpath" ] && { echo "[ERROR] Unable to find VM using VM UUID:" $vmuuid ; exit $ERROR_VC_OBJECT_NOT_FOUND; }
+ERROR_MSG="Unable to find VM using VM UUID: ${vmuuid}"
+[ -z "$vmpath" ] && {update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"; exit $ERROR_VC_OBJECT_NOT_FOUND; }
 
 govc vm.change -e="disk.enableUUID=1" -vm="$vmpath" &> /dev/null
 if [ $? -eq 0 ]; then
     echo "[INFO] Successfully enabled disk.enableUUID flag on the Node Virtual Machine".
 else
-    echo "[ERROR] Failed to enable disk.enableUUID flag on the Node Virtual Machine"
+    ERROR_MSG="Failed to enable disk.enableUUID flag on the Node Virtual Machine"
+    update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
     exit $ERROR_ENABLE_DISK_UUID
 fi
+
+PHASE=$DAEMONSET_SCRIPT_PHASE3
+update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_RUNNING" ""
 
 # Move Node VM to the VM Folder.
 govc object.mv $vmpath $k8s_secret_node_vms_folder &> /dev/null
 if [ $? -eq 0 ]; then
     echo "[INFO] Moved Node Virtual Machine to the Working Directory Folder".
 else
-    echo "[ERROR] Failed to move Node Virtual Machine to the Working Directory Folder"
+    ERROR_MSG="Failed to move Node Virtual Machine to the Working Directory Folder"
+    update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
     exit $ERROR_MOVE_NODE_TO_WORKING_DIR
 fi
+
+
+PHASE=$DAEMONSET_SCRIPT_PHASE4
+update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_RUNNING" ""
 
 # Creating back up directory for manifest files and kubelet service configuration file.
 echo "[INFO] Creating directory: '/host/tmp/$POD_NAME' for back up of manifest files and kubelet service configuration file"
@@ -50,7 +72,8 @@ mkdir /host/tmp/$POD_NAME
 if [ $? -eq 0 ]; then
     echo "[INFO] Successfully created back up directory:" /host/tmp/$POD_NAME " on $NODE_NAME node"
 else
-    echo "[ERROR] Failed to create directory: '/host/tmp/$POD_NAME' for back up of manifest files and kubelet service configuration file"
+    ERROR_MSG="Failed to create directory: '/host/tmp/${POD_NAME}' for back up of manifest files and kubelet service configuration file"
+    update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
     exit $ERROR_FAIL_TO_CREATE_BACKUP_DIRECTORY
 fi
 
@@ -59,7 +82,8 @@ ls /host/$k8s_secret_vcp_configuration_file_location &> /dev/null
 if [ $? -eq 0 ]; then
     echo "[INFO] Verified that the directory for the vSphere Cloud Provider configuration file is accessible. Path: ("/host/$k8s_secret_vcp_configuration_file_location ")"
 else
-    echo "[ERROR] Directory (/host/$k8s_secret_vcp_configuration_file_location) for vSphere Cloud Provider Configuration file is not present"
+    ERROR_MSG="Directory (/host/${k8s_secret_vcp_configuration_file_location}) for vSphere Cloud Provider Configuration file is not present"
+    update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
     exit $ERROR_VSPHERE_CONF_DIRECTORY_NOT_PRESENT
 fi
 
@@ -70,7 +94,8 @@ if [ $? -eq 0 ]; then
     if [ $? -eq 0 ]; then
         echo "[INFO] Existing vsphere.conf file is copied to" /host/tmp/$POD_NAME/vsphere.conf
     else
-        echo "[ERROR] Failed to back up vsphere.conf file at " /host/tmp/$POD_NAME/vsphere.conf
+        ERROR_MSG="Failed to back up vsphere.conf file at " /host/tmp/${POD_NAME}/vsphere.conf
+        update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
         exit $ERROR_FAIL_TO_BACKUP_FILE
     fi
 fi
@@ -78,13 +103,17 @@ fi
 # locate and back up manifest files and kubelet service configuration file.
 file=/host/$k8s_secret_kubernetes_api_server_manifest
 backupdir=/host/tmp/$POD_NAME
-locate_validate_and_backup_files $file $backupdir
+locate_validate_and_backup_files $file $backupdir $POD_NAME
 
 file=/host/$k8s_secret_kubernetes_controller_manager_manifest
-locate_validate_and_backup_files $file $backupdir
+locate_validate_and_backup_files $file $backupdir $POD_NAME
 
 file=/host/$k8s_secret_kubernetes_kubelet_service_configuration_file
-locate_validate_and_backup_files $file $backupdir
+locate_validate_and_backup_files $file $backupdir $POD_NAME
+
+
+PHASE=$DAEMONSET_SCRIPT_PHASE5
+update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_RUNNING" ""
 
 # Create vSphere Cloud Provider configuration file
 echo "[INFO] Creating vSphere Cloud Provider configuration file at /host/tmp/vsphere.conf"
@@ -103,9 +132,13 @@ echo "[Global]
 if [ $? -eq 0 ]; then
     echo "[INFO] successfully created vSphere.conf file at :" /host/tmp/vsphere.conf
 else
-    echo "[ERROR] Failed to create vsphere.conf file at :" /host/tmp/vsphere.conf
+    ERROR_MSG="Failed to create vsphere.conf file at : /host/tmp/vsphere.conf"
+    update_VcpConigStatus "$POD_NAME" "$PHASE" "FAILED" "$ERROR_MSG"
     exit $ERROR_FAIL_TO_CREATE_FILE
 fi
+
+PHASE=$DAEMONSET_SCRIPT_PHASE6
+update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_RUNNING" ""
 
 # update manifest files
 ls /host/$k8s_secret_kubernetes_api_server_manifest &> /dev/null
@@ -114,19 +147,30 @@ if [ $? -eq 0 ]; then
     if [ "${k8s_secret_kubernetes_api_server_manifest##*.}" == "json" ]; then
         MANIFEST_FILE="/host/tmp/kube-apiserver.json"
         cp /host/$k8s_secret_kubernetes_api_server_manifest $MANIFEST_FILE
-        add_flags_to_manifest_file $MANIFEST_FILE
+        add_flags_to_manifest_file $MANIFEST_FILE $POD_NAME
     elif [ "${k8s_secret_kubernetes_api_server_manifest##*.}" == "yaml" ]; then
         YAML_MANIFEST_FILE="/host/tmp/kube-apiserver.yaml"
         JSON_MANIFEST_FILE="/host/tmp/kube-apiserver.json"
         cp /host/$k8s_secret_kubernetes_api_server_manifest $YAML_MANIFEST_FILE
         # Convert YAML to JSON format
         j2y -r $YAML_MANIFEST_FILE > $JSON_MANIFEST_FILE
-        add_flags_to_manifest_file $JSON_MANIFEST_FILE
+        if [ $? -ne 0 ]; then
+            ERROR_MSG="Failed to convert file from YAML to JSON format"
+            update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
+            exit $ERROR_J2Y_FAILURE
+        fi
+        add_flags_to_manifest_file $JSON_MANIFEST_FILE $POD_NAME
         # Convert JSON to YAML foramt
         j2y $JSON_MANIFEST_FILE > $YAML_MANIFEST_FILE
+        if [ $? -ne 0 ]; then
+            ERROR_MSG="Failed to convert file from JSON to YAML format"
+            update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
+            exit $ERROR_J2Y_FAILURE
+        fi
         rm -rf $JSON_MANIFEST_FILE
     else
-        echo "[ERROR] Unsupported file format"
+        ERROR_MSG="Unsupported file format"
+        update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
         exit $ERROR_UNSUPPORTED_FILE_FORMAT
     fi
 fi
@@ -137,7 +181,7 @@ if [ $? -eq 0 ]; then
     if [ "${k8s_secret_kubernetes_controller_manager_manifest##*.}" == "json" ]; then
         MANIFEST_FILE="/host/tmp/kube-controller-manager.json"
         cp /host/$k8s_secret_kubernetes_controller_manager_manifest $MANIFEST_FILE
-        add_flags_to_manifest_file $MANIFEST_FILE
+        add_flags_to_manifest_file $MANIFEST_FILE $POD_NAME
     elif [ "${k8s_secret_kubernetes_controller_manager_manifest##*.}" == "yaml" ]; then
         YAML_MANIFEST_FILE="/host/tmp/kube-controller-manager.yaml"
         JSON_MANIFEST_FILE="/host/tmp/kube-controller-manager.json"
@@ -145,19 +189,22 @@ if [ $? -eq 0 ]; then
         # Convert YAML to JSON format
         j2y -r $YAML_MANIFEST_FILE > $JSON_MANIFEST_FILE
         if [ $? -ne 0 ]; then
-            echo "[ERROR] Failed to convert file from YAML to JSON format"
+            ERROR_MSG="Failed to convert file from YAML to JSON format"
+            update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
             exit $ERROR_J2Y_FAILURE
         fi
-        add_flags_to_manifest_file $JSON_MANIFEST_FILE
+        add_flags_to_manifest_file $JSON_MANIFEST_FILE $POD_NAME
         # Convert JSON to YAML foramt
         j2y $JSON_MANIFEST_FILE > $YAML_MANIFEST_FILE
         if [ $? -ne 0 ]; then
-            echo "[ERROR] Failed to convert file from JSON to YAML format"
+            ERROR_MSG="Failed to convert file from JSON to YAML format"
+            update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
             exit $ERROR_J2Y_FAILURE
         fi
         rm -rf $JSON_MANIFEST_FILE
     else
-        echo "[ERROR] Unsupported file format"
+        ERROR_MSG="Unsupported file format"
+        update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
         exit $ERROR_UNSUPPORTED_FILE_FORMAT
     fi
 fi
@@ -184,7 +231,8 @@ if [ $? -eq 0 ]; then
     if [ $? -eq 0 ]; then 
         echo "[INFO] Sucessfully updated kubelet.service configuration"
     else
-        echo "[ERROR] Failed to update kubelet.service configuration"
+        ERROR_MSG="Failed to update kubelet.service configuration"
+        update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
         exit $ERROR_FAIL_TO_ADD_CONFIG_PARAMETER
     fi
 fi
@@ -214,6 +262,9 @@ if [ -f /host/tmp/kubelet-service-configuration ]; then
     cp /host/tmp/kubelet-service-configuration /host/$k8s_secret_kubernetes_kubelet_service_configuration_file
 fi
 
+PHASE=$DAEMONSET_SCRIPT_PHASE7
+update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_RUNNING" ""
+
 echo "systemctl daemon-reload
 systemctl restart ${k8s_secret_kubernetes_kubelet_service_name}
 " > /host/tmp/restart_kubelet.sh
@@ -224,5 +275,8 @@ chroot /host /tmp/restart_kubelet.sh
 if [ $? -eq 0 ]; then
     echo "[INFO] kubelet service restarted sucessfully"
 else
-    echo "[ERROR] Failed to restart kubelet service"
+    ERROR_MSG="Failed to restart kubelet service"
+    update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_FAILED" "$ERROR_MSG"
 fi
+
+update_VcpConigStatus "$POD_NAME" "$PHASE" "$DAEMONSET_PHASE_COMPLETE" ""
